requirements.txt

Flask
requests
dnspython
